+++ libs/core/user-menu/directives/user-menu-control-element.directive.ts
@@ -1,0 +1,7 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+    // eslint-disable-next-line @angular-eslint/directive-selector
+    selector: '[fd-user-menu-control-element]'
+})
+export class UserMenuControlElementDirective {}